fix theme toggler
