+++ src/components/Layout/Layout.js
@@ -5,26 +5,29 @@ import Header from '../Header/Header'
 import Footer from '../Footer/Footer'
 import { rhythm } from '../../utils/typography'
 import { GlobalStyles } from '../GlobalStyles/GlobalStyles'
-import { ThemeModeProvider } from '../ThemeProvider/ThemeProvider'
+import { ThemeModeProvider } from '../useThemeMode'
+import { ThemeProvider } from '../ThemeProvider/ThemeProvider'
 
 const Content = styled.div`
   max-width: ${rhythm(25)};
   margin: 0 auto;
-  padding: ${rhythm(0.5)}};
+  padding: ${rhythm(0.5)};
 `
 
 const Layout = ({ children }) => (
   <ThemeModeProvider>
-    <Fragment>
-      <Helmet
-        titleTemplate="%s · RayGesualdo.com"
-        defaultTitle="RayGesualdo.com"
-      />
-      <GlobalStyles />
-      <Header />
-      <Content>{children}</Content>
-      <Footer />
-    </Fragment>
+    <ThemeProvider>
+      <Fragment>
+        <Helmet
+          titleTemplate="%s · RayGesualdo.com"
+          defaultTitle="RayGesualdo.com"
+        />
+        <GlobalStyles />
+        <Header />
+        <Content>{children}</Content>
+        <Footer />
+      </Fragment>
+    </ThemeProvider>
   </ThemeModeProvider>
 )
